feat: add basic heartbeat
